move getPersonalInfoAPI to utils.ts
